# services/rota/core/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A high-performance proxy rotation server with health monitoring and intelligent routing\nProvides comprehensive API for managing proxy servers, monitoring their health,\nand configuring rotation strategies.",
        "title": "Rota Proxy API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/alpkeskin/rota"
        },
        "license": {
            "name": "LICENSE",
            "url": "https://github.com/alpkeskin/rota/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8001",
    "basePath": "/api/v1",
    "paths": {
...
        "/settings": {
            "get": {
                "description": "Get current system configuration",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Get settings",
                "responses": {
                    "200": {
                        "description": "Current settings",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.Settings"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update system configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Update settings",
                "parameters": [
                    {
                        "description": "Updated settings",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.Settings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update confirmation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/settings/geoip/status": {
            "get": {
                "description": "Get the effective GeoIP provider, active database, and updater state",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Get GeoIP database status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.GeoIPStatus"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/settings/geoip/update-db": {
            "post": {
                "description": "Download, validate, atomically replace, and activate the configured MaxMind database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Update managed GeoIP database",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/settings/reset": {
            "post": {
                "description": "Reset configuration to default values",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Reset settings",
                "responses": {
                    "200": {
                        "description": "Reset confirmation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_alpkeskin_rota_core_internal_models.AuthenticationSettings": {
            "type": "object",
            "properties": {
                "enabled": {
                    "description": "Enable authentication for proxy requests",
                    "type": "boolean"
                },
                "password": {
                    "description": "Password for proxy authentication (write-only)",
                    "type": "string"
                },
                "username": {
                    "description": "Username for proxy authentication",
                    "type": "string"
                }
            }
        },
...
        "github_com_alpkeskin_rota_core_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                }
            }
        },
        "github_com_alpkeskin_rota_core_internal_models.GeoIPSettings": {
            "type": "object",
            "properties": {
                "auto_update": {
                    "type": "boolean"
                },
                "last_updated_at": {
                    "type": "string"
                },
                "maxmind_db_path": {
                    "type": "string"
                },
                "maxmind_license_key": {
                    "description": "Write-only.",
                    "type": "string"
                },
                "maxmind_url": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "update_interval_hours": {
                    "type": "integer"
                }
            }
        },
        "github_com_alpkeskin_rota_core_internal_models.GeoIPStatus": {
            "type": "object",
            "properties": {
                "active_database_path": {
                    "type": "string"
                },
                "auto_update": {
                    "type": "boolean"
                },
                "configured": {
                    "type": "boolean"
                },
                "database_build_time": {
                    "type": "string"
                },
                "database_loaded": {
                    "type": "boolean"
                },
                "database_path": {
                    "type": "string"
                },
                "database_type": {
                    "type": "string"
                },
                "last_error": {
                    "type": "string"
                },
                "last_updated_at": {
                    "type": "string"
                },
                "license_configured": {
                    "type": "boolean"
                },
                "provider": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "update_interval_hours": {
                    "type": "integer"
                },
                "updating": {
                    "type": "boolean"
                }
            }
        },
        "github_com_alpkeskin_rota_core_internal_models.HealthCheckSettings": {
            "type": "object",
            "properties": {
                "base_status": {
                    "type": "integer"
                },
                "base_url": {
                    "type": "string"
                },
                "headers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "integer"
                },
                "strict_tls": {
                    "type": "boolean"
                },
                "timeout": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                },
                "workers": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_alpkeskin_rota_core_internal_models.LogRetentionSettings": {
            "type": "object",
            "properties": {
                "cleanup_interval_hours": {
                    "description": "How often to run cleanup (1, 6, 12, 24)",
                    "type": "integer"
                },
                "compression_after_days": {
                    "description": "Compress logs older than X days (1, 3, 7, 14)",
                    "type": "integer"
                },
                "enabled": {
                    "description": "Enable automatic log cleanup",
                    "type": "boolean"
                },
                "retention_days": {
                    "description": "Days to keep logs (7, 15, 30, 60, 90)",
                    "type": "integer"
                }
            }
        },
...
        "github_com_alpkeskin_rota_core_internal_models.ProxyCleanupSettings": {
            "type": "object",
            "properties": {
                "cleanup_interval_hours": {
                    "description": "How often to run cleanup",
                    "type": "integer"
                },
                "enabled": {
                    "description": "Enable automatic dead proxy cleanup",
                    "type": "boolean"
                },
                "max_failed_days": {
                    "description": "Remove proxies failed for more than N days",
                    "type": "integer"
                },
                "min_success_rate": {
                    "description": "Remove proxies with success rate below X% (0 = disabled)",
                    "type": "number"
                }
            }
        },
        "github_com_alpkeskin_rota_core_internal_models.ProxyLifecycleSettings": {
            "type": "object",
            "properties": {
                "active_recheck_minutes": {
                    "type": "integer"
                },
                "auto_archive_enabled": {
                    "type": "boolean"
                },
                "hard_unreachable_after_hours": {
                    "type": "integer"
                },
                "soft_unreachable_after_hours": {
                    "type": "integer"
                },
                "youtube_unusable_after_hours": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_alpkeskin_rota_core_internal_models.RateLimitSettings": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "interval": {
                    "description": "in seconds",
                    "type": "integer"
                },
                "max_requests": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_alpkeskin_rota_core_internal_models.RotationSettings": {
            "type": "object",
            "properties": {
                "allowed_protocols": {
                    "description": "Empty means all supported protocols.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fallback": {
                    "type": "boolean"
                },
                "fallback_max_retries": {
                    "type": "integer"
                },
                "follow_redirect": {
                    "type": "boolean"
                },
                "max_response_time": {
                    "description": "in milliseconds, 0 means no limit",
                    "type": "integer"
                },
                "method": {
                    "type": "string"
                },
                "min_success_rate": {
                    "description": "0-100, 0 means no minimum",
                    "type": "number"
                },
                "remove_unhealthy": {
                    "type": "boolean"
                },
                "retries": {
                    "type": "integer"
                },
                "time_based": {
                    "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.TimeBasedSettings"
                },
                "timeout": {
                    "type": "integer"
                }
            }
        },
        "github_com_alpkeskin_rota_core_internal_models.Settings": {
            "type": "object",
            "properties": {
                "authentication": {
                    "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.AuthenticationSettings"
                },
                "geoip": {
                    "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.GeoIPSettings"
                },
                "healthcheck": {
                    "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.HealthCheckSettings"
                },
                "log_retention": {
                    "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.LogRetentionSettings"
                },
                "proxy_cleanup": {
                    "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ProxyCleanupSettings"
                },
                "proxy_lifecycle": {
                    "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ProxyLifecycleSettings"
                },
                "rate_limit": {
                    "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.RateLimitSettings"
                },
                "rotation": {
                    "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.RotationSettings"
                }
            }
        },
...
        "github_com_alpkeskin_rota_core_internal_models.TimeBasedSettings": {
            "type": "object",
            "properties": {
                "interval": {
                    "description": "in seconds",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}